tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-app-history-back-forward-same-doc.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = t.step_timeout(resolve, 0));
  await appHistory.navigate("#foo");
  assert_equals(appHistory.entries().length, 2);

  let oncurrentchange_back_called = false;
  let back_committed = false;
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_back_called = true;
    assert_equals(e.from, appHistory.entries()[1]);
    assert_equals(e.navigationType, "traverse");
    assert_equals(appHistory.current.index, 0);
    assert_false(back_committed);
  });
  let back_result = appHistory.back();
  assert_false(oncurrentchange_back_called);
  await back_result.committed.then(() => back_committed = true);
  assert_true(oncurrentchange_back_called);

  let oncurrentchange_forward_called = false;
  let forward_committed = false;
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_forward_called = true;
    assert_equals(e.from, appHistory.entries()[0]);
    assert_equals(e.navigationType, "traverse");
    assert_equals(appHistory.current.index, 1);
    assert_false(forward_committed);
  });
  let forward_result = appHistory.forward();
  assert_false(oncurrentchange_forward_called);
  await forward_result.committed.then(() => forward_committed = true);
  assert_true(oncurrentchange_forward_called);
}, "AppHistoryCurrentChangeEvent fires for appHistory.back() and appHistory.forward()");
</script>
